Proper InertiaCancelling=false in config handling
In my testing of disabling the config I realized that it still would just tell the user that the Toggle was activated/deactivated. Added better handling that will always force-disable in item NBT when a user tries to toggle but feature is disabled on server. A special message letting the user know it's disabled on the server will display when they try to toggle the feature with keybind.